[LibArchFPGA] Moved CAD Types out of Lib
The CAD types header file does not belong in libarchfpga since it was
only really used in the prepacker of VPR. It makes several references to
the pb_graph structure which I would ultimately like to move into VPR
entirely.
