+++ vpr/src/pack/prepack.h
@@ -9,8 +9,9 @@
  */
 
 #include <algorithm>
+#include <map>
 #include "atom_netlist_fwd.h"
-#include "cad_types.h"
+#include "pack_patterns.h"
 #include "vtr_assert.h"
 #include "vtr_range.h"
 #include "vtr_strong_id.h"
@@ -21,6 +22,7 @@
 class t_pack_molecule;
 class LogicalModels;
 struct t_logical_block_type;
+struct t_pb_graph_node;
 
 // A unique ID used to identify a molecule generated by the prepacker.
 typedef vtr::StrongId<struct pack_molecule_id_tag, size_t> PackMoleculeId;
